initial gene record refactor
