+++ apps/genomicsdb/app/globals.css
@@ -49,6 +49,7 @@
 
     /* Layout */
     --header-height: 72px;
+    --footer-height: 90px;
     --sidebar-width: 280px;
     --border-radius: 8px;
     --border-radius-sm: 4px;
