+++ apps/genomicsdb/components/records/record-sidebar.css
@@ -144,14 +144,20 @@
 
 .record-page-container {
     display: flex;
-    min-height: calc(100vh - var(--header-height));
+    min-height: calc(100vh - var(--header-height) - var(--footer-height));
     margin-top: 0;
+}
+
+.record-page-container.capped {
+    height: calc(100vh - var(--header-height));
 }
 
 .record-content {
     flex: 1;
     padding: 0 2rem;
     transition: margin-left var(--transition-normal);
+
+    overflow-y: scroll;
 }
 
 .record-content.with-sidebar {
