+++ apps/genomicsdb/lib/client-wrapper.ts
@@ -1,0 +1,3 @@
+"use client";
+
+export * from "@niagads/ui/client";
