+++ packages/UI/src/styles/footer.css
@@ -3,10 +3,12 @@
     border-top: 1px solid var(--gray-800);
     margin-top: auto;
     color: var(--gray-300);
+    height: var(--footer-height)
 }
 
 .footer-container {
     max-width: 1200px;
+    height: 100%;
     margin: 0 auto;
     padding: 1.5rem 2rem;
     display: flex;
